invest_app/requirements.txt
MetaTrader5>=5.0.45
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.25.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
pytest>=7.4.0
